use custom __m128 struct on non-win32 platforms to provide element-wise access explicitly cast param for set_constant to correctly resolve overloaded function

+++ src/alloy/backend/ivm/ivm_assembler.cc
@@ -81,7 +81,7 @@ int IVMAssembler::Assemble(
     size_t type_size = GetTypeSize(slot->type);
     // Align to natural size.
     stack_offset = XEALIGN(stack_offset, type_size);
-    slot->set_constant(stack_offset);
+    slot->set_constant((uint32_t)stack_offset);
     stack_offset += type_size;
   }
   // Ensure 16b alignment.
